Undo change to method visibility
Upstream code would not be able to adopt this change because it is
backwards-incompatible.  This would be the resulting situation:

<?php
class A {
    public function test() {
        echo "A::test()\n";
    }
}
class B extends A {
    protected function test() {
        echo "B::test()\n";
        parent::test();
    }

    public function call_test() {
        $this->test();
    }
}
$b = new B();
$b->call_test();
?>

Fatal error: Access level to B::test() must be public (as in class A).

The reverse situation (overriding a protected method and changing its
visibility to public) is allowed, so that's what this commit does
instead.  See also:  https://stackoverflow.com/q/16877875