Do not require `jetpack_require_lib()` to exist on the site
If `class WPCom_GHF_Markdown_Parser` already exists, then we can just
use it.

This allows site owners to integrate Markdown sync without requiring all
of Jetpack.  In this case they should copy the relevant code into their
site from here, and load it themselves:

https://github.com/Automattic/jetpack/tree/master/_inc/lib/markdown

As of this writing this code works independently of the rest of Jetpack.