Create table-of-contents.php
This is code taken from Automattic/Nacin's Handbook plugin (which covers many things besides a Toc), simplified, and with enhancements for accessibility added.

One question (which I have included in line 25 of the code) is whether we want this to create a ToC on all pages (not just those with code examples). I think it might be a good idea to have it everywhere, but I'm interested in others' views. If so, we won't need the conditional check at lines 25-28.